fix bug;
1. EstimateFeeRate接口中RPC节点返回fee单位是btc, 需要改为sat单位; 修改后要同步删除freeapis的临时处理代码;
2. 将mempool返回余额数字类型从int改为float

+++ btcapis_address.go
@@ -14,7 +14,7 @@ func (c *Client) ImportAddressAndPublickey(ctx context.Context, address string, 
 }
 
 // GetAddressBalance 返回地址的确认余额和未确认余额.
-func (c *Client) GetAddressBalance(ctx context.Context, addr string) (confirmed int64, mempool int64, err error) {
+func (c *Client) GetAddressBalance(ctx context.Context, addr string) (confirmed float64, mempool float64, err error) {
 	return c.addressClient.GetAddressBalance(ctx, addr)
 }
