+++ internal/address/apis.go
@@ -9,7 +9,7 @@ import (
 )
 
 // GetAddressBalance 通过地址, 获取地址的确认余额和未确认余额.
-func (c *Client) GetAddressBalance(ctx context.Context, addr string) (confirmed int64, mempool int64, err error) {
+func (c *Client) GetAddressBalance(ctx context.Context, addr string) (confirmed float64, mempool float64, err error) {
 	if c.mempoolapisClient != nil {
 		return c.mempoolapisClient.AddressGetBalance(ctx, addr)
 	}
